+++ app/src/test/kotlin/com/coinepro/app/ScreenshotRenderTest.kt
@@ -701,7 +701,7 @@ class ScreenshotRenderTest {
 
     @Test
     fun tools() = capture("05-tools") {
-        ToolsScreen(onOpenConnections = {}, onOpenNews = {}, onOpenCalendar = {}, onOpenJournal = {})
+        ToolsScreen(onOpenConnections = {}, onOpenNews = {}, onOpenCalendar = {}, onOpenJournal = {}, onOpenPaperTrade = {})
     }
 
     @Test
@@ -768,7 +768,7 @@ class ScreenshotRenderTest {
     @Test
     @Config(sdk = [34], qualifiers = "fa-rIR-ldrtl-w411dp-h914dp-xxhdpi")
     fun toolsPersian() = capture("67-tools-fa") {
-        ToolsScreen(onOpenConnections = {}, onOpenNews = {}, onOpenCalendar = {}, onOpenJournal = {})
+        ToolsScreen(onOpenConnections = {}, onOpenNews = {}, onOpenCalendar = {}, onOpenJournal = {}, onOpenPaperTrade = {})
     }
 
     @Test